refactor : 동적 html 생성 관리 클래스 구현 - HtmlManager를 통해서 HtmlRoot를 생성하도록 변경하였습니다.
