+++ src/main/java/codesquad/handler/TemplateResourceHandler.java
@@ -6,22 +6,20 @@ import codesquad.http.ContentType;
 import codesquad.http.HttpRequest;
 import codesquad.http.HttpResponse;
 import codesquad.http.HttpStatus;
-import codesquad.template.HtmlElement;
+import codesquad.template.HtmlManager;
 import codesquad.template.HtmlRoot;
 import codesquad.template.Model;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.util.Deque;
-import java.util.LinkedList;
 
 public class TemplateResourceHandler implements HttpHandler {
 
     private final Logger logger = LoggerFactory.getLogger(StaticResourceHandler.class);
+    private final HtmlManager htmlManager = new HtmlManager();
 
     @Override
     public boolean match(HttpRequest request) {
@@ -50,46 +48,11 @@ public class TemplateResourceHandler implements HttpHandler {
         };
     }
 
-    private void writeFileToBody(String path, HttpResponse response) {
-
-        try (InputStream in = this.getClass().getResourceAsStream("/templates" + path)) {
-            byte[] fileContentBytes = in.readAllBytes();
-            response.setBody(fileContentBytes);
-        } catch (FileNotFoundException e) {
-            logger.error("File not found: {}", path);
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-            logger.error("Error reading from binary file: {}", path);
-            throw new RuntimeException(e);
-        }
-    }
-
     public void convertToHtml(String path, HttpRequest request, HttpResponse response) {
         try (InputStream in = this.getClass().getResourceAsStream("/templates" + path)) {
             String template = new String(in.readAllBytes());
-            HtmlRoot root = new HtmlRoot();
-            // split tags
-            Deque<String> tags = new LinkedList<>();
-            StringBuilder tag = new StringBuilder();
-            char[] templates = template.toCharArray();
-            for (int i = 0; i < template.length(); i++) {
-                if (templates[i] == '<') {
-                    if (!tag.isEmpty() && !tag.toString().trim().isEmpty()) tags.add(tag.toString());
-                    tag = new StringBuilder();
-                    tag.append(templates[i]);
-                } else if (templates[i] == '>') {
-                    tag.append(templates[i]);
-                    if (!tag.isEmpty() && !tag.toString().trim().isEmpty()) tags.add(tag.toString());
-                    tag = new StringBuilder();
-                } else {
-                    tag.append(templates[i]);
-                }
-            }
 
-            while (!tags.isEmpty()) {
-                HtmlElement.Builder elementBuilder = extractElement(tags);
-                root.addElement(elementBuilder.build());
-            }
+            HtmlRoot root = htmlManager.create(template);
 
             SessionContext session = SessionContextManager.getSession(request);
 
@@ -102,32 +65,5 @@ public class TemplateResourceHandler implements HttpHandler {
             logger.error("Error reading from binary file: {}", path);
             throw new RuntimeException(e);
         }
-    }
-
-    private HtmlElement.Builder extractElement(Deque<String> tags) {
-        String firstTags = tags.pollFirst();
-        HtmlElement.Builder elementBuilder = HtmlElement.create(firstTags);
-
-        if (!elementBuilder.isOpen()) {
-            return elementBuilder;
-        }
-
-        String nextTag = tags.peekFirst();
-
-        while (!tags.isEmpty()) {
-            nextTag = tags.peekFirst();
-            if (!elementBuilder.isOpen()) {
-                break;
-            }
-            if (nextTag.contains("/" + elementBuilder.getTag())) {
-                tags.pollFirst();
-                break;
-            }
-            elementBuilder.addChildren(extractElement(tags));
-        }
-        if (nextTag != null) {
-            elementBuilder.setClose();
-        }
-        return elementBuilder;
     }
 }
